requirements.txt
Django==3.1.5
django-cors-headers==3.7.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
gunicorn==20.0.4
mysql-connector-python==8.0.23
PyMySQL==1.0.2
python-decouple==3.4